Feature/role

+++ app/Controllers/Http/RolesController.ts
@@ -2,8 +2,6 @@ import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import { ModelPaginatorContract } from '@ioc:Adonis/Lucid/Orm'
 import Role from 'App/Models/Role'
 import RoleValidator from 'App/Validators/RoleValidator'
-import Database from '@ioc:Adonis/Lucid/Database'
-import { DateTime } from 'luxon'
 import Permission from 'App/Models/Permission'
 import User from 'App/Models/User'
 import CommonFilterQueryValidator from 'App/Validators/CommonFilterQueryValidator'
@@ -41,28 +39,16 @@ export default class RolesController {
     return await RoleService.getPaginatedRoles(ctx, false)
   }
 
-  public async store({ bouncer, request, response }: HttpContextContract): Promise<void> {
-    await bouncer.with('RolePolicy').authorize('create')
-    await request.validate(RoleValidator)
-    const role = new Role()
-    const trx = await Database.transaction()
+  public async store(ctx: HttpContextContract): Promise<void> {
+    await ctx.bouncer.with('RolePolicy').authorize('create')
+    await ctx.request.validate(RoleValidator)
+    const role = await RoleService.saveRole(ctx, new Role())
 
-    role.name = request.input('name')
-    role.slug = request.input('slug')
-    role.description = request.input('description', undefined)
-
-    try {
-      await role.useTransaction(trx).save()
-      await role.related('permissions').attach(request.input('permissions'), trx)
-      await trx.commit()
-    } catch (error) {
-      await trx.rollback()
-      return response.internalServerError({
-        message: 'Something went wrong. Please try again.',
-      })
-    }
-
-    return response.created(role)
+    return role
+      ? ctx.response.created(role)
+      : ctx.response.internalServerError({
+          message: 'Something went wrong. Please try again later.',
+        })
   }
 
   public async show({ bouncer, params }: HttpContextContract): Promise<Role> {
@@ -74,46 +60,36 @@ export default class RolesController {
     await ctx.bouncer.with('RolePolicy').authorize('update')
     const role = await Role.findByOrFail('slug', ctx.params.slug)
     await ctx.request.validate(new RoleValidator(ctx, role.id))
-    const trx = await Database.transaction()
+    const res = await RoleService.saveRole(ctx, role)
 
-    role.name = ctx.request.input('name')
-    role.slug = ctx.request.input('slug')
-    role.description = ctx.request.input('description')
-
-    try {
-      await role.useTransaction(trx).save()
-      await role.related('permissions').sync(ctx.request.input('permissions'), undefined, trx)
-      await trx.commit()
-    } catch (error) {
-      await trx.rollback()
-      return ctx.response.internalServerError({
-        message: 'Something went wrong. Please try again.',
-      })
-    }
-
-    return ctx.response.noContent()
+    return res
+      ? ctx.response.noContent()
+      : ctx.response.internalServerError({
+          message: 'Something went wrong. Please try again later.',
+        })
   }
 
   public async destroy({ bouncer, params, response }: HttpContextContract): Promise<void> {
     await bouncer.with('RolePolicy').authorize('delete')
     const role = await Role.findByOrFail('slug', params.slug)
+
+    if (role.id === 1) {
+      return response.badRequest({ message: 'Operation not permitted' })
+    }
+
     await role.delete()
     return response.noContent()
   }
 
   public async deactivate({ bouncer, params, response }: HttpContextContract): Promise<void> {
     await bouncer.with('RolePolicy').authorize('activate')
-    const role = await Role.findByOrFail('slug', params.slug)
-    role.deactivatedAt = DateTime.now()
-    await role.save()
-    return response.noContent()
+    const res = await RoleService.updateState(params.slug, false)
+    return res ? response.noContent() : response.badRequest({ message: 'Operation not permitted' })
   }
 
   public async activate({ bouncer, params, response }: HttpContextContract): Promise<void> {
     await bouncer.with('RolePolicy').authorize('activate')
-    const role = await Role.findByOrFail('slug', params.slug)
-    role.deactivatedAt = undefined
-    await role.save()
+    await RoleService.updateState(params.slug, true)
     return response.noContent()
   }
 
@@ -124,22 +100,27 @@ export default class RolesController {
     return role.permissions
   }
 
-  public async users({
-    bouncer,
-    params,
-    request,
-  }: HttpContextContract): Promise<ModelPaginatorContract<User>> {
-    await bouncer.with('RolePolicy').authorize('viewUsers')
-    const role = await Role.findByOrFail('slug', params.slug)
+  public async users(ctx: HttpContextContract): Promise<ModelPaginatorContract<User>> {
+    await ctx.bouncer.with('RolePolicy').authorize('viewUsers')
+    const role = await Role.findByOrFail('slug', ctx.params.slug)
+    await ctx.request.validate(
+      new CommonFilterQueryValidator(ctx, ['name', 'email', 'created_at', 'updated_at'])
+    )
 
-    const page = request.input('page', 1)
-    const perPage = request.input('perPage', 10)
-    const orderBy = request.input('orderBy', 'name')
-    const orderDirection = request.input('orderDirection', 'asc')
+    const page = ctx.request.input('page', 1)
+    const perPage = ctx.request.input('perPage', 10)
+    const search = ctx.request.input('query')
+    const sortBy = ctx.request.input('sortBy', 'name')
+    const order = ctx.request.input('order', 'asc')
 
-    return await User.query()
-      .whereHas('roles', (query) => query.where('id', role.id))
-      .orderBy(orderBy, orderDirection)
-      .paginate(page, perPage)
+    const query = role.related('users').query().orderBy(sortBy, order)
+
+    if (search) {
+      query.andWhere((q) => {
+        q.where('name', 'like', `%${search}%`).orWhere('name', 'like', `%${search}%`)
+      })
+    }
+
+    return query.paginate(page, perPage)
   }
 }
